requirements.txt
groq
requests
streamlit
python-dotenv
pandas
plotly
streamlit-mic-recorder
SpeechRecognition
pyaudio